Revert "LPD-30592 Move LanguageState to bundle's top level package"
This reverts commit 6002ef16af161c92344af0a64162cddaf02b866d.
