+++ modules/apps/frontend-js/frontend-js-web/src/main/java/com/liferay/frontend/js/web/internal/language/LanguageState.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.frontend.js.web.internal;
+package com.liferay.frontend.js.web.internal.language;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.language.Language;
