+++ modules/apps/frontend-js/frontend-js-web/src/main/java/com/liferay/frontend/js/web/internal/model/listener/PLOEntryModelListener.java
@@ -5,7 +5,7 @@
 
 package com.liferay.frontend.js.web.internal.model.listener;
 
-import com.liferay.frontend.js.web.internal.LanguageState;
+import com.liferay.frontend.js.web.internal.language.LanguageState;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.log.Log;
