+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/module-log4j.xml
@@ -2,7 +2,7 @@
 
 <Configuration strict="true">
 	<Loggers>
-		<Logger level="ERROR" name="com.liferay.frontend.js.web.internal.LanguageState" />
+		<Logger level="ERROR" name="com.liferay.frontend.js.web.internal.language.LanguageState" />
 		<Logger level="ERROR" name="com.liferay.frontend.js.web.internal.model.listener.PLOEntryModelListener" />
 		<Logger level="ERROR" name="com.liferay.frontend.js.web.internal.servlet.FrontendJsWebLanguageServlet" />
 		<Logger level="ERROR" name="com.liferay.frontend.js.web.internal.servlet.ResourceBundleServiceTrackerCustomizer" />
